2014-09-05 22:03 UTC+0200 Viktor Szakats (vszakats users.noreply.github.com)   * contrib/gtwvg/tests/_activex.prg     * cleanups
  * src/rtl/val.c
    * whitespace formatting
    % unnecessary cast

+++ src/rtl/val.c
@@ -61,7 +61,7 @@ HB_FUNC( VAL )
       HB_MAXINT lValue;
       double dValue;
 
-      fDbl = hb_valStrnToNum( szText, iLen, &lValue, &dValue , &iDec, &iWidth );
+      fDbl = hb_valStrnToNum( szText, iLen, &lValue, &dValue, &iDec, &iWidth );
 
       if( ! fDbl )
          hb_retnintlen( lValue, iWidth );
@@ -84,10 +84,10 @@ HB_FUNC( HB_VAL )
       HB_MAXINT lValue;
       double dValue;
 
-      fDbl = hb_valStrnToNum( szText, iLen, &lValue, &dValue , &iDec, &iWidth );
+      fDbl = hb_valStrnToNum( szText, iLen, &lValue, &dValue, &iDec, &iWidth );
 
       if( HB_ISNUM( 2 ) )
-         iLen = ( int ) hb_parni( 2 );
+         iLen = hb_parni( 2 );
 
       if( fDbl && iDec > 0 )
          iLen -= iDec + 1;
